Take into account invalid params/payload from express body-parser
JSON body-parser generates errors when non-JSON values are POSTed as params/payload to HTTP endpoints.
We should not send back a 500 when these types of errors are raised and should send back a 400 instead.
The body of the error should have the same structure as those that arise from swagger validation failures.